+++ ArcadeGame1/src/arcadegame1package1/GameMain.java
@@ -129,7 +129,7 @@ public class GameMain extends Canvas implements Runnable{
 	}
 	
 	// This method makes sure the player stays within the game boundaries
-	public static int clamp(int var, int min, int max){
+	public static float clamp(float var, float min, float max){
 		if(var >= max)
 			return var = max;
 		else if (var <= min)
