+++ ArcadeGame1/src/arcadegame1package1/HUD.java
@@ -5,11 +5,11 @@ import java.awt.Graphics;
 
 public class HUD {
 
-	public static int Health = 100;
+	public static float Health = 100;
 	
-	private int greenValue = 255;
-	private int score = 0;
-	private int level = 1;
+	private float greenValue = 255;
+	private float score = 0;
+	private float level = 1;
 	
 	public void tick(){
 		Health = GameMain.clamp(Health, 0, 100);
@@ -23,27 +23,27 @@ public class HUD {
 	public void render(Graphics g){
 		g.setColor(Color.white);
 		g.fillRect(15, 15, 200, 32);
-		g.setColor(new Color(75, greenValue, 0));
-		g.fillRect(15, 15, Health * 2, 32);
+		g.setColor(new Color(75, (int)greenValue, 0));
+		g.fillRect((int)15, (int)15, (int)Health * 2, 32);
 		g.setColor(Color.white);
 		g.drawRect(15, 15, 200, 32);
 		g.drawString("Score: " + score, 15, 64);
 		g.drawString("Level: " + level, 15, 80);
 	}
 	
-	public void score(int score){
+	public void score(float score){
 		this.score = score;
 	}
 	
-	public int getScore(){
+	public float getScore(){
 		return score;
 	}
 	
-	public int getLevel(){
+	public float getLevel(){
 		return level;
 	}
 	
-	public void setLevel(int level){
+	public void setLevel(float level){
 		this.level = level;
 	}
 }
